+++ packages/playground-ui/src/ds/components/PropertyFilter/index.ts
@@ -1,4 +1,4 @@
-export * from './property-filter';
+export * from './types';
 export * from './property-filter-actions';
 export * from './property-filter-applied';
 export * from './property-filter-creator';
